+++ backend/src/packages/chaiNNer_external/external_stable_diffusion/automatic1111/txt2img.py
@@ -10,7 +10,6 @@ from nodes.impl.external_stable_diffusion import (
     STABLE_DIFFUSION_TEXT2IMG_PATH,
     SamplerName,
     decode_base64_image,
-    nearest_valid_size,
     post,
     verify_api_connection,
 )
@@ -24,7 +23,7 @@ from nodes.properties.inputs import (
 )
 from nodes.properties.outputs import ImageOutput
 from nodes.utils.seed import Seed
-from nodes.utils.utils import get_h_w_c
+from nodes.utils.utils import get_h_w_c, nearest_valid_size
 
 from .. import auto1111_group
