+++ backend/src/packages/chaiNNer_stable_diffusion/stable_diffusion/latent/empty_latent_image.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import torch
 
-from nodes.impl.external_stable_diffusion import nearest_valid_size
+from nodes.utils.utils import nearest_valid_size
 from nodes.impl.stable_diffusion import LatentImage
 from nodes.properties.inputs import SliderInput
 from nodes.properties.outputs.stable_diffusion_outputs import LatentImageOutput
