+++ backend/src/packages/chaiNNer_stable_diffusion/stable_diffusion/latent/latent_upscale.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import torch
 
-from nodes.impl.external_stable_diffusion import nearest_valid_size
+from nodes.utils.utils import nearest_valid_size
 from nodes.impl.stable_diffusion import CropMethod, LatentImage, UpscaleMethod
 from nodes.properties.inputs import EnumInput, SliderInput
 from nodes.properties.inputs.stable_diffusion_inputs import LatentImageInput
